Move 2 text helper methods out into helper module
