+++ codegen/codegen.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative 'generators/text_helpers'
+
 require_relative 'generators/base'
 
 require_relative 'generators/cpp'
